+++ tests/e2e/README.md
@@ -1,5 +1,7 @@
 ## End To End Testing
 
+## TODO documentation here
+
 E2e tests are categorized into files as follows:
 
 - `setup_test.go` - setup for the e2e tests
