+++ tests/e2e/slashing_test.go
@@ -1,4 +1,4 @@
-package e2e_test
+package e2e
 
 import (
 	"fmt"
