+++ tests/e2e/stop_consumer_test.go
@@ -1,4 +1,4 @@
-package e2e_test
+package e2e
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
