+++ tests/e2e/unbonding_test.go
@@ -1,4 +1,4 @@
-package e2e_test
+package e2e
 
 import (
 	"time"
